+++ pages/cv.tsx
@@ -4,7 +4,7 @@ import Metadata from '../source/components/metadata';
 import CvEducationSection from '../source/components/cv/educationSection';
 import CvEmploymentSection from '../source/components/cv/employmentSection';
 import CvPortfolioSection from '../source/components/cv/portfolioSection';
-import CvProfileSection from '../source/components/cvProfileSection';
+import CvProfileSection from '../source/components/cv/profileSection';
 import CvSkillsSection from '../source/components/cvSkillsSection';
 
 import Education from '../source/types/education';
